Call Realtime

+++ client/src/components/message/RightSide.js
@@ -12,7 +12,7 @@ import LoadIcon from '../../images/loading.gif'
 
 
 const RightSide = () => {
-    const { auth, message, theme, socket } = useSelector(state => state)
+    const { auth, message, theme, socket, peer } = useSelector(state => state)
     const dispatch = useDispatch()
     
     const { id } = useParams()
@@ -138,8 +138,46 @@ const RightSide = () => {
     }, [isLoadMore])
 
     const handleDeleteConversation = () => {
-        dispatch(deleteConversation({auth, id}))
-        return history.push('/message')
+        if(window.confirm('Do you want to delete?')){
+            dispatch(deleteConversation({auth, id}))
+            return history.push('/message')
+        }
+    }
+
+    // Call
+    const caller = ({video}) => {
+        const { _id, avatar, username, fullname } = user
+
+        const msg = {
+            sender: auth.user._id,
+            recipient: _id, 
+            avatar, username, fullname, video
+        }
+        dispatch({ type: GLOBALTYPES.CALL, payload: msg })
+    }
+
+    const callUser = ({video}) => {
+        const { _id, avatar, username, fullname } = auth.user
+
+        const msg = {
+            sender: _id,
+            recipient: user._id, 
+            avatar, username, fullname, video
+        }
+
+        if(peer.open) msg.peerId = peer._id
+
+        socket.emit('callUser', msg)
+    }
+
+    const handleAudioCall = () => {
+        caller({video: false})
+        callUser({video: false})
+    }
+    
+    const handleVideoCall = () => {
+        caller({video: true})
+        callUser({video: true})
     }
 
     return (
@@ -148,8 +186,16 @@ const RightSide = () => {
                 {
                     user.length !== 0 &&
                     <UserCard user={user}>
-                        <i className="fas fa-trash text-danger"
-                        onClick={handleDeleteConversation} />
+                        <div>
+                            <i className="fas fa-phone-alt"
+                            onClick={handleAudioCall} />
+
+                            <i className="fas fa-video mx-3"
+                            onClick={handleVideoCall} />
+
+                            <i className="fas fa-trash text-danger"
+                            onClick={handleDeleteConversation} />
+                        </div>
                     </UserCard>
                 }
             </div>
